router: split classes in separate files
